+++ src/main/java/org/socketio/netty/session/AbstractSocketSession.java
@@ -24,21 +24,20 @@ public abstract class AbstractSocketSession extends AbstractSession {
 	private final Channel channel;
 
 	public AbstractSocketSession(
-			TransportType transport,
 			Channel channel, 
 			String sessionId, 
 			String origin, 
 			ISessionDisconnectHandler disconnectHandler, 
 			final TransportType upgradedFromTransportType, 
 			int localPort) {
-		super(transport, channel, sessionId, origin, disconnectHandler, upgradedFromTransportType, localPort);
+		super(channel, sessionId, origin, disconnectHandler, upgradedFromTransportType, localPort);
 		this.channel = channel;
 	}
 	
 	@Override
 	public void send(Packet packet) {
 		if (packet != null && channel != null && channel.isConnected()) {
-			preparePacket(packet);
+			fillPacketHeaders(packet);
 			channel.write(packet);
 		}
 	}