+++ src/main/java/org/socketio/netty/session/FlashSocketSession.java
@@ -21,7 +21,7 @@ import org.socketio.netty.TransportType;
 public class FlashSocketSession extends AbstractSocketSession {
 	
 	public FlashSocketSession(Channel channel, String sessionId, String origin, ISessionDisconnectHandler disconnectHandler, final TransportType upgradedFromTransportType, int localPort) {
-		super(TransportType.FLASHSOCKET, channel, sessionId, origin, disconnectHandler, upgradedFromTransportType, localPort);
+		super(channel, sessionId, origin, disconnectHandler, upgradedFromTransportType, localPort);
 	}
 
 	@Override